Refactored the code to be more clear
