+++ common/src/main/java/software/bluelib/markdown/syntax/Strikethrough.java
@@ -1,7 +1,8 @@
 // Copyright (c) BlueLib. Licensed under the MIT License.
 
-package software.bluelib.utils.markdown;
+package software.bluelib.markdown.syntax;
 
+import software.bluelib.markdown.MarkdownFeature;
 import software.bluelib.utils.logging.BaseLogLevel;
 import software.bluelib.utils.logging.BaseLogger;
