+++ neoforge/src/main/java/software/bluelib/example/event/ChatHandler.java
@@ -6,7 +6,7 @@ import net.minecraft.network.chat.Component;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.event.ServerChatEvent;
-import software.bluelib.utils.markdown.MarkdownParser;
+import software.bluelib.markdown.MarkdownParser;
 
 /**
  * A {@code public class} responsible for handling server chat events and formatting chat messages using Markdown.
